Move exceptions tests to test_exceptions (#560)
